rename __gvectors, __aux, __readfiles removing the __
