+++ irrep/aux.py
